Fixed strange compiler error
Reference to assertEquals is ambiguous. It is defined in void assertEquals(java.lang.String, long, long) and void assertEquals(java.lang.String, int, int).

by upcasting a few ints to longs

(Not sure if this came up during the upgrade to JUnit 3.7 or because I'm using a 1.3 compiler with different ambiguity rules.)

+++ test/src/java/org/jdom/contrib/xpath/test/XPath_Test.java
@@ -96,7 +96,7 @@ public class XPath_Test extends XPathTestCase {
     Element childG = relem(XML1).getChild("g");
     Comment comment = new XPathElement(childG).getComment();
     List matches = match(childG, "child::comment()");
-    assertEquals("size", 1, matches.size());
+    assertEquals("size", (long)1, matches.size());
     assertEquals("match the same object", comment, matches.get(0));
   }
 
@@ -105,7 +105,7 @@ public class XPath_Test extends XPathTestCase {
     Element childG = relem(XML1).getChild("g");
     List piList = new XPathElement(childG).getProcessingInstructions();
     List matches = match(childG, "child::processing-instruction()");
-    assertEquals("size", 2, matches.size());
+    assertEquals("size", (long)2, matches.size());
     assertEquals("match the same object", piList.get(0), matches.get(0));
     assertEquals("match the same object", piList.get(1), matches.get(1));
   }
@@ -115,7 +115,7 @@ public class XPath_Test extends XPathTestCase {
     Element childG = relem(XML1).getChild("g");
     List piList = new XPathElement(childG).getProcessingInstructions("gpi1");
     List matches = match(childG, "child::processing-instruction('gpi1')");
-    assertEquals("size", 1, matches.size());
+    assertEquals("size", (long)1, matches.size());
     assertEquals("match the same object", piList.get(0), matches.get(0));
   }
